+++ apps/duskmoon_bundler/test/mix/tasks/duskmoon_bundler.install_test.exs
@@ -122,6 +122,8 @@ defmodule DuskmoonBundler.InstallTest do
         igniter.rewrite.sources["mix.exs"]
         |> Rewrite.Source.get(:content)
 
+      assert mix_content =~ ~s({:duskmoon_bundler_runtime, "~> 9.7"})
+      assert mix_content =~ "{:duskmoon_bundler, \"~> 9.7\", runtime: Mix.env() == :dev}"
       assert mix_content =~ ~s("assets.setup": [])
       assert mix_content =~ ~s("assets.build": ["compile", "duskmoon_bundler.build --tailwind"])